requirements.txt
Django==1.11
django-material
requests
requests-toolbelt